hardly fix SortByKeys

+++ utils/collection/sort_test.go
@@ -3,13 +3,13 @@ package collection
 import "testing"
 
 func TestSortByKeys(t *testing.T) {
-	m := map[int]string{
+	m := map[int64]string{
 		1: "a",
 		3: "b",
 		2: "c",
 		0: "d",
 		5: "e",
 	}
-	sorted := SortByKeys[int](m)
+	sorted := SortByKeys[int64](m)
 	logger.Println(sorted)
 }
